+++ saltapi/auth/authorization.py
@@ -12,36 +12,7 @@ from starlette.requests import HTTPConnection
 
 from saltapi.auth.token import parse_token
 from saltapi.repository import user_repository
-from saltapi.repository.user_repository import User, is_user_pc, is_user_pi, is_user_verified
-
-
-class Permission(enum.Enum):
-    """A permission."""
-
-    SUBMIT_PROPOSAL = "SUBMIT_PROPOSAL"
-    VIEW_PROPOSAL = "VIEW_PROPOSAL"
-
-    @staticmethod
-    def from_name(name: str) -> "Permission":
-        """Return the permission for a name."""
-        for _name, member in Permission.__members__.items():
-            if _name == name:
-                return member
-        raise ValueError(f"Unknown permission: {name}")
-
-
-class Role(enum.Enum):
-    """A role."""
-
-    ADMINISTRATOR = "ADMINISTRATOR"
-
-    @staticmethod
-    def from_name(name: str) -> "Role":
-        """Return the permission for a name."""
-        for _name, member in Role.__members__.items():
-            if _name == name:
-                return member
-        raise ValueError(f"Unknown role: {name}")
+from saltapi.repository.user_repository import User, is_user_pc, is_user_pi, Permission
 
 
 class AuthenticatedUser(BaseUser):
@@ -102,14 +73,7 @@ class TokenAuthenticationBackend(AuthenticationBackend):
         return AuthCredentials(["authenticated"]), AuthenticatedUser(user)
 
 
-def has_permission(user: User, permission: Permission) -> bool:
-    """Check whether the user has a role."""
-    if permission in user.permissions:
-        return True
-    return False
-
-
-def can_submit_proposal(user: User, proposal_code: Optional[str] = None) -> None:
+def can_submit_proposal(user: User, proposal_code: Optional[str] = None) -> bool:
     """
     Check whether a user can submit a proposal.
 
@@ -126,12 +90,13 @@ def can_submit_proposal(user: User, proposal_code: Optional[str] = None) -> None
         If the the user can re submit the proposal.
 
     """
-    if not is_user_verified:
-        raise Exception("User not verified.8")
+
+    if not user.has_permission(Permission.SUBMIT_PROPOSAL):
+        return False
     if not proposal_code:  # Is a new proposal
-        return
+        return True
     if is_user_pi(user.username, proposal_code):  # User is a principal investigator
-        return
+        return True
     if is_user_pc(user.username, proposal_code):  # User is a principal contact
-        return
-    raise Exception("Only the Principal Investigator and Principal Contact may resubmit a proposal.")
+        return True
+    return False
