+++ saltapi/graphql/schema.graphql
@@ -25,7 +25,7 @@ For example, consider the following type:
 ```
 type Proposal {
   content(proposalCode: ProposalCode): String!
-                @permittedFor(roles=[PROPOSAL_OWNER], permissions=[VIEW_ALL_PROPOSALS])
+                @isAuthenticated(roles=[PROPOSAL_OWNER], permissions=[VIEW_ALL_PROPOSALS])
 }
 ```
 
@@ -34,7 +34,7 @@ if they may view all proposals.
 
 The directive uses the arguments of the field resolver as required.
 """
-directive @permittedFor(roles: [Role!], permissions: [Permission!]) on FIELD_DEFINITION
+directive @isAuthenticated on FIELD_DEFINITION
 
 type Query {
     """
@@ -62,7 +62,7 @@ type Mutation {
         The proposal code of the proposal to which this block belongs.
         """
         proposalCode: ProposalCode!
-    ): ID
+    ): ID @isAuthenticated
 
     """
     Submit a proposal.
@@ -87,7 +87,7 @@ type Mutation {
         The proposal code.
         """
         proposalCode: ProposalCode
-    ): ID
+    ): ID @isAuthenticated
 }
 
 type Subscription {
