+++ tests/test_authentication.py
@@ -23,6 +23,7 @@ async def mock_find_user_by_credentials(username: str, password: str) -> Optiona
             email="jane@example.com",
             roles=[],
             permissions=[],
+            is_authenticated=True
         )
     else:
         return None
@@ -39,6 +40,7 @@ async def mock_find_user_by_id(user_id: int) -> Optional[User]:
             email="jane@example.com",
             roles=[],
             permissions=[],
+            is_authenticated=True
         )
     else:
         return None
